Move preimage templates to folder
