+++ packages/next-common/components/preImages/createPreimagePopup/templateButtons.js
@@ -1,6 +1,4 @@
 import { ChoiceButton } from "next-common/components/summary/newProposalButton/common";
-import { useChainSettings } from "next-common/context/chain";
-import { useState } from "react";
 import { InfoPopoular } from "@osn/icons/subsquare";
 
 export function FellowshipTreasurySpendButton({ onClick }) {
@@ -72,124 +70,4 @@ export function KillReferendumButton({ onClick }) {
       onClick={onClick}
     />
   );
-}
-
-export function useFellowshipTreasurySpendButton() {
-  const [
-    showFellowshipTreasurySpendPopup,
-    setShowFellowshipTreasurySpendPopup,
-  ] = useState(false);
-
-  const fellowshipTreasurySpendButton = (
-    <FellowshipTreasurySpendButton
-      onClick={() => setShowFellowshipTreasurySpendPopup(true)}
-    />
-  );
-
-  return {
-    fellowshipTreasurySpendButton,
-    showFellowshipTreasurySpendPopup,
-  };
-}
-
-export function useSpendLocalTreasuryButton() {
-  const [showSpendLocalTreasuryPopup, setShowSpendLocalTreasuryPopup] =
-    useState(false);
-
-  const localTreasuryButton = (
-    <SpendLocalTreasuryButton
-      onClick={() => setShowSpendLocalTreasuryPopup(true)}
-    />
-  );
-
-  return {
-    localTreasuryButton,
-    showSpendLocalTreasuryPopup,
-  };
-}
-
-export function useSpendUSDxTreasuryButton() {
-  const [showSpendUSDxTreasuryPopup, setShowSpendUSDxTreasuryPopup] =
-    useState(false);
-  const {
-    treasuryProposalTracks,
-    newProposalQuickStart: { usdxTreasuryProposal } = {},
-  } = useChainSettings();
-
-  const usdxTreasuryButton = treasuryProposalTracks && usdxTreasuryProposal && (
-    <SpendUSDxTreasuryButton
-      onClick={() => setShowSpendUSDxTreasuryPopup(true)}
-    />
-  );
-
-  return {
-    usdxTreasuryButton,
-    showSpendUSDxTreasuryPopup,
-  };
-}
-
-export function useSpendDotOnAssetHubButton() {
-  const [showSpendDotOnAssetHubPopup, setShowSpendDotOnAssetHubPopup] =
-    useState(false);
-  const {
-    treasuryProposalTracks,
-    newProposalQuickStart: { spendDotOnAssetHubProposal } = {},
-  } = useChainSettings();
-
-  const spendDotOnAssetHubButton = treasuryProposalTracks &&
-    spendDotOnAssetHubProposal && (
-      <SpendDotOnAssetHubButton
-        onClick={() => setShowSpendDotOnAssetHubPopup(true)}
-      />
-    );
-
-  return {
-    spendDotOnAssetHubButton,
-    showSpendDotOnAssetHubPopup,
-  };
-}
-
-export function useNewRemarkButton() {
-  const [showNewRemarkPopup, setShowNewRemarkPopup] = useState(false);
-
-  const remarkButton = (
-    <NewRemarkButton onClick={() => setShowNewRemarkPopup(true)} />
-  );
-
-  return {
-    remarkButton,
-    showNewRemarkPopup,
-  };
-}
-
-export function useCancelReferendumButton() {
-  const [showCancelReferendumPopup, setShowCancelReferendumPopup] =
-    useState(false);
-  const { newProposalQuickStart: { cancelReferendum } = {} } =
-    useChainSettings();
-
-  const cancelReferendumButton = cancelReferendum && (
-    <CancelReferendumButton
-      onClick={() => setShowCancelReferendumPopup(true)}
-    />
-  );
-
-  return {
-    cancelReferendumButton,
-    showCancelReferendumPopup,
-  };
-}
-
-export function useKillReferendumButton() {
-  const [showKillReferendumPopup, setShowKillReferendumPopup] = useState(false);
-  const { newProposalQuickStart: { killReferendum } = {} } = useChainSettings();
-
-  const killReferendumButton = killReferendum && (
-    <KillReferendumButton onClick={() => setShowKillReferendumPopup(true)} />
-  );
-
-  return {
-    killReferendumButton,
-    showKillReferendumPopup,
-  };
 }
