+++ packages/next-common/components/preImages/createPreimagePopup/templates/newLocalTreasuryProposalPopup.js
@@ -5,9 +5,9 @@ import { useContextApi } from "next-common/context/api";
 import { useChainSettings } from "next-common/context/chain";
 import { checkInputValue } from "next-common/utils";
 import Popup from "next-common/components/popup/wrapper/Popup";
-import NotePreimageButton from "./notePreimageButton";
-import useBalanceField from "./fields/useBalanceField";
-import useAddressComboField from "./fields/useAddressComboField";
+import NotePreimageButton from "../notePreimageButton";
+import useBalanceField from "../fields/useBalanceField";
+import useAddressComboField from "../fields/useAddressComboField";
 import { usePopupParams } from "next-common/components/popupWithSigner/context";
 
 export function useLocalTreasuryNotePreimageTx(inputBalance, beneficiary) {
